Split template functions into a separate file
